services/auth-service/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.2.1
httpx==0.28.1
pydantic-settings==2.13.1
email-validator==2.3.0
